make-release: make mtime, owner, group consistent
Files from git have modification time set to one
of commit, and owner/group to root.
Making it so for generated files as well makes
it easier to generate an identical tarball from git.

Setting owner/group to root is especially important because
otherwise you must have a user/group with same name
to generate an identical tarball.

Signed-off-by: Michael S. Tsirkin <mst@redhat.com>
Signed-off-by: Marcelo Tosatti <mtosatti@redhat.com>