device-assignment: Don't deassign when the assignment fails
The last thing assign_device() does is call into KVM_ASSIGN_PCI_DEVICE.
If that fails, the device is not assigned, so we shouldn't then try to
deassign it.  If you try to assign the same device multiple times, you
can get into a nasty fail-succeed-fail-succeed loop.  And we certainly
shouldn't take the assigned_out branch before we've even attempted
to assign the device.

Signed-off-by: Alex Williamson <alex.williamson@redhat.com>
Signed-off-by: Marcelo Tosatti <mtosatti@redhat.com>