device-assignment: be more selective in interrupt disabling
An 82576 physical function assigned to a Windows 7 guest currently
doesn't work because the driver seems to gratuitiously disable
MSI and MSIX interrupts.  When it does this, we blindly deassign
the current interrupt setup, leaving the device with no interrupts.
Instead let's only deassign the irq if we were previously using
MSI/MSIX or if we're going to start using MSI/MSIX.

Signed-off-by: Alex Williamson <alex.williamson@redhat.com>
Acked-by: Chris Wright <chrisw@redhat.com>
Signed-off-by: Avi Kivity <avi@redhat.com>