make-release: fix mtime on rhel6 beta
With recent git versions (like one in rhel6 beta),
git prints out tag info in addition to
requested format, if passed a tag name.
This breaks make-release.
To get just the time as we want, we need to
pass something that is not a tag name but matches
same commit.  $commit^{commit} does this.

Signed-off-by: Michael S. Tsirkin <mst@redhat.com>
Signed-off-by: Marcelo Tosatti <mtosatti@redhat.com>