device-assignment: Byte-wise ROM read
The host kernel filters the PCI option ROM, returning only bytes for
the actual ROM size, not for the whole BAR.  That means we typically
do a short read of the PCI sysfs ROM file.  Read it a byte at a time
so we know how much to actually copy and only skip the copy if we
get nothing.

Signed-off-by: Alex Williamson <alex.williamson@redhat.com>
Acked-by: Chris Wright <chrisw@redhat.com>
Signed-off-by: Avi Kivity <avi@redhat.com>