kvm: drop kvm_context parameter from msix-related kvm functions
The kvm_context parameter serves no useful purpose since it's a global (and
all callers reference that global); on the other hand it prevents the build
when kvm is not configured in since the kvm_context symbol is not defined.

Remove the parameter and replace it by the global kvm_context in callees.

Acked-by: Michael S. Tsirkin <mst@redhat.com>
Signed-off-by: Avi Kivity <avi@redhat.com>