device-assignment: handle device with incorrect PCIe Cap structure size
The bcm5761 provides a PCIe Cap structure (capid=0x10)
that is invalid, providing one that is 8 bytes shorter
than the v2 PCIe spec defines.
This leads to a memory corruption when mapped for device-assigment.

Add a check in assigned_device_pci_cap_init() to correct
this hw error for this device, and try to catch other ones
and print warnings if they exists.

Signed-off-by: Donald Dutile <ddutile@redhat.com>
Acked-by: Alex Williamson <alex.williamson@redhat.com>
Acked-by: Michael S. Tsirking <mst@redhat.com>
Signed-off-by: Avi Kivity <avi@redhat.com>